+++ src/models/Settings.php
@@ -34,11 +34,13 @@ class Settings extends Model
 
     /**
      * @var string|null The OAuth token.
+     * @deprecated in 2.1.5
      */
     public $token;
 
     /**
      * @var string|null The OAuth token secret.
+     * @deprecated in 2.1.5
      */
     public $tokenSecret;